Рефакторинг render-comments
